addons/GodotPilot/python/requirements.txt
flask==3.1.0
grpcio==1.70.0
grpcio-tools==1.70.0
langgraph==0.2.74
langchain==0.3.19
langchain-google-genai==2.0.9
langchain-ollama==0.2.3
langchain-openai==0.3.6
python-dotenv==1.0.1
